listeners: Treat binding errors as critical failures
When trying to listen on a socket, and binding the configured address fails,
treat the error as critical.  Systemd will then restart the process after a
while.  The failure typically happens because an IP address was not yet
available, and ignoring the error may result in a partially working service.
Restart may fix it once the address is configured.